For clients that use sprockets.mixins.metrics, DB metrics use a "db" subpath.
For example, if a program calls self.postgres_execute() with
metric_name='insert-user', that will record a metric on path
"db.insert-user".

This is particularly beneficial for visualizing metrics.
By querying Graphite for all metrics that have "db" in their path,
you can easily visualize all DB metrics for a program.
If someone later adds another DB query and metric to that program,
it meets the same conditions,
thus it can be automatically added to the graph.